Revert "rename tawhiri.api.v1 module to tawhiri.api"
This reverts commit 78fe78fb734aed9d9a0a8721c479dc93c5f2d39a.

Conflicts:
	tawhiri/api/__init__.py
	tawhiri/app.py
	tests/testapi.py
