+++ tawhiri/app.py
@@ -21,7 +21,7 @@ The main Tawhiri web application
 
 from flask import Flask
 
-from .api import blueprint as api_v1
+from .api.v1 import blueprint as api_v1
 
 app = Flask(__name__)
 app.register_blueprint(api_v1, url_prefix='/api/v1')
