+++ tests/testapi.py
@@ -7,7 +7,7 @@ from flask.ext.testing import TestCase
 from mock import patch, MagicMock
 from six.moves.urllib.parse import urlencode
 
-from tawhiri.api import blueprint
+from tawhiri.api.v1 import blueprint
 
 # Root path for v1 API
 API_ROOT = '/api/v1/'
@@ -37,8 +37,8 @@ class BasicApiTest(TestCase):
 
     @patch('tawhiri.models.standard_profile')
     @patch('tawhiri.solver.solve')
-    @patch('tawhiri.api.WindDataset')
-    @patch('tawhiri.api.ruaumoko_ds')
+    @patch('tawhiri.api.v1.WindDataset')
+    @patch('tawhiri.api.v1.ruaumoko_ds')
     def test_simple_run(self, ruaumoko_ds_mock, wind_ds_mock, solve_mock, profile_mock):
         """Make a simple request for a landing prediction."""
